Refactor y Fix de UI
